local_agents/function_calling/tool_simulation/requirements.txt
absl-py
google-ai-generativelanguage
google-generativeai
protobuf
proto-plus
tensorflow